requirements.txt
pyttsx3==2.90
SpeechRecognition==3.10.0
pyaudio==0.2.13
wikipedia==1.4.0
pyautogui==0.9.54
pywhatkit==5.4
requests==2.31.0
pyperclip==1.8.2
nltk==3.8.1
opencv-python==4.8.1.78
numpy==1.26.4
python-dotenv
newsapi-python
groq>=0.4.0
pywin32==306